port varfont_info.py script into a gftools subcommand
(issue #3)
